+++ exporter/exporterhelper/internal/persistent_storage_test.go
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build enable_unstable
-// +build enable_unstable
-
 package internal
 
 import (
@@ -49,7 +46,9 @@ func createTestClient(extension storage.Extension) storage.Client {
 }
 
 func createTestPersistentStorageWithLoggingAndCapacity(client storage.Client, logger *zap.Logger, capacity uint64) *persistentContiguousStorage {
-	return newPersistentContiguousStorage(context.Background(), "foo", capacity, logger, client, newFakeTracesRequestUnmarshalerFunc())
+	pcs := newPersistentContiguousStorage("foo", capacity, logger, newFakeTracesRequestUnmarshalerFunc())
+	pcs.start(context.Background(), client)
+	return pcs
 }
 
 func createTestPersistentStorage(client storage.Client) *persistentContiguousStorage {