+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### 💡 Enhancements 💡
 
+- Enable Persistent Queue by default. (#5711)
 - `ocb` now exits with an error if it fails to load the build configuration. (#5731)
 - Deprecate `HTTPClientSettings.ToClientWithHost` (#5737)